+++ starters/node-typescript/src/modules/identity/public/permissions.ts
@@ -1,4 +1,5 @@
 export const Permissions = {
   UserCreate: 'identity.user.create',
   UserRead: 'identity.user.read',
+  Login: 'identity.login',
 } as const;